+++ BackEnd/Routes/itemsRoutes.js
@@ -1,32 +1,9 @@
 const express = require('express');
 const router = express.Router();
 
-const multer = require('multer');
-const path = require('path')
-
-
-const uploadDirectory = path.join(__dirname, 'uploads');
-
-const storage = multer.diskStorage({
-    destination: function (req, file, cb) {
-        cb(null, uploadDirectory);
-    },
-    filename: function (req, file, cb) {
-        const uniqueSuffix = Date.now() + '-' + Math.round(Math.random() * 1E9);
-        cb(null, uniqueSuffix + '-' + file.originalname);
-    }
-});
-
-const upload = multer({ storage: storage });
-
-// Use 'upload.fields' to handle multiple files with different field names
-const uploadMiddleware = upload.fields([
-    { name: 'ItemFeaturedPhoto', maxCount: 1 },
-    { name: 'ItemOtherPhotos', maxCount: 5 }, // Adjust maxCount as needed
-]);
-
-
 const Controllers = require('../Controllers')
+// import the upload middleware
+const multerMiddleware = require('../Middleware/multerMiddleware');
 
 // route to get all items from the DB for the itemcards:
 router.get('/getall', (req, res) => {
@@ -43,10 +20,12 @@ router.get('/', (res) => {
 });
 
 // route to create an item in the database (createpostmodal):
-router.post('/create', uploadMiddleware, (req, res) => {
+router.post('/create', multerMiddleware([
+    { name: 'ItemFeaturedPhoto', maxCount: 1 },
+    { name: 'ItemOtherPhotos', maxCount: 5 },
+]), (req, res) => {
     console.log('Request Body:', req.body); // testing
     console.log('Request Files:', req.files);  // testing
-    // Call the controller function to create the item
     Controllers.itemController.createItem(req, res);
 });
